requirements.txt
streamlit
pandas
numpy
yfinance
tensorflow
scikit-learn
matplotlib
alpha_vantage
python-dotenv
plotly
